working configuration, updated other things

+++ src/main/java/com/zti/partpicker/controller/CPUController.java
@@ -1,12 +1,14 @@
 package com.zti.partpicker.controller;
 
-import java.util.List;
-
 import com.zti.partpicker.exception.CPUNotFoundException;
 import com.zti.partpicker.model.CPU;
 import com.zti.partpicker.repository.CPURepository;
+import com.zti.partpicker.repository.ConfigurationRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @CrossOrigin(origins = "http://localhost:3000")
 @RestController
@@ -16,8 +18,12 @@ public class CPUController {
     @Autowired
     private final CPURepository repository;
 
-    CPUController(CPURepository repository) {
+    @Autowired
+    private final ConfigurationRepository configurationRepository;
+
+    CPUController(CPURepository repository, ConfigurationRepository configurationRepository) {
         this.repository = repository;
+        this.configurationRepository = configurationRepository;
     }
 
     @GetMapping
@@ -37,7 +43,12 @@ public class CPUController {
     }
 
     @DeleteMapping("/{id}")
-    void deleteCPU(@PathVariable Long id) {
+    ResponseEntity<String> deleteCPU(@PathVariable Long id) {
+        if (configurationRepository.findAllByCpu(id).size() > 0) {
+            return ResponseEntity.badRequest().build();
+        };
+
         repository.deleteById(id);
+        return ResponseEntity.ok().build();
     }
 }
