+++ src/main/java/com/zti/partpicker/controller/GPUController.java
@@ -4,9 +4,11 @@ import java.util.List;
 
 import com.zti.partpicker.exception.GPUNotFoundException;
 import com.zti.partpicker.model.GPU;
+import com.zti.partpicker.repository.ConfigurationRepository;
 import com.zti.partpicker.repository.GPURepository;
 import jakarta.annotation.security.RolesAllowed;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @CrossOrigin(origins = "http://localhost:3000")
@@ -17,8 +19,12 @@ public class GPUController {
     @Autowired
     private final GPURepository repository;
 
-    GPUController(GPURepository repository) {
+    @Autowired
+    private final ConfigurationRepository configurationRepository;
+
+    GPUController(GPURepository repository, ConfigurationRepository configurationRepository) {
         this.repository = repository;
+        this.configurationRepository = configurationRepository;
     }
 
     @RolesAllowed("ROLE_ADMIN")
@@ -40,7 +46,12 @@ public class GPUController {
     }
 
     @DeleteMapping("/{id}")
-    void deleteGPU(@PathVariable Long id) {
+    ResponseEntity<String> deleteGPU(@PathVariable Long id) {
+        if (configurationRepository.findAllByGpu(id).size() > 0) {
+            return ResponseEntity.badRequest().build();
+        };
+
         repository.deleteById(id);
+        return ResponseEntity.ok().build();
     }
 }
