+++ src/main/java/com/zti/partpicker/controller/MemoryController.java
@@ -4,8 +4,10 @@ import java.util.List;
 
 import com.zti.partpicker.exception.MemoryNotFoundException;
 import com.zti.partpicker.model.Memory;
+import com.zti.partpicker.repository.ConfigurationRepository;
 import com.zti.partpicker.repository.MemoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @CrossOrigin(origins = "http://localhost:3000")
@@ -16,8 +18,12 @@ public class MemoryController {
     @Autowired
     private final MemoryRepository repository;
 
-    MemoryController(MemoryRepository repository) {
+    @Autowired
+    private final ConfigurationRepository configurationRepository;
+
+    MemoryController(MemoryRepository repository, ConfigurationRepository configurationRepository) {
         this.repository = repository;
+        this.configurationRepository = configurationRepository;
     }
 
     @GetMapping
@@ -37,7 +43,12 @@ public class MemoryController {
     }
 
     @DeleteMapping("/{id}")
-    void deleteMemory(@PathVariable Long id) {
+    ResponseEntity<String> deleteMemory(@PathVariable Long id) {
+        if (configurationRepository.findAllByMemory(id).size() > 0) {
+            return ResponseEntity.badRequest().build();
+        };
+
         repository.deleteById(id);
+        return ResponseEntity.ok().build();
     }
 }
